requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
matplotlib==3.7.2
seaborn==0.12.2
streamlit==1.28.0
fastapi==0.103.0
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6
openai==1.3.0
langchain==0.0.305
pinecone-client==2.2.3
sentence-transformers==2.2.2
joblib==1.3.2
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0
scipy==1.11.2
PyPDF2==3.0.1
python-docx==1.1.0
